requirements.txt
imageio==2.23.0
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.17.0
tensorflow==2.10.0
tensorflow_gpu==2.10.0
